feat/webrtc: Wire up to SAFE
Instead of requesting the user to copy around the startup info, we now store it via the SAFE-launcher on the network itself. Another user with the same room id can then retrieve that information and they can both connect. Note: this requires the new LOW_LEVEL_API which isn't fully featured yet, so it provides a mock-launcher. It is also not yet super clean in that it doesn't delete it's own connection data from the network nor closes the handlers