feat/comments: Implement updating of comment
The controller now save a given comment as a new version on the structured data entry
saving the original time (for sorting) and adding the timestamp of the edit. Further
the view shows that date and does minor checks on the save of the edit to ensure we
don't waste network capacity.